Refactor package, mostly by renaming and combining modules
